// frontend/src/api/generated/models/log-fragment-rarity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * ログの欠片のレアリティ
 * @export
 * @enum {string}
 */

export const LogFragmentRarity = {
  Common: 'common',
  Uncommon: 'uncommon',
  Rare: 'rare',
  Epic: 'epic',
  Legendary: 'legendary',
  Unique: 'unique',
  Architect: 'architect',
} as const

export type LogFragmentRarity =
  (typeof LogFragmentRarity)[keyof typeof LogFragmentRarity]